chore: Convert motor speed values to integers for consistency

+++ app.py
@@ -26,7 +26,7 @@ def motor_control(position_x):
     global motor_speeds
     motor_speed_initial = 150
     max_speed_diff = 50
-    motor_speeds = [motor_speed_initial + position_x * max_speed_diff, motor_speed_initial - position_x * max_speed_diff, motor_speed_initial - position_x * max_speed_diff, motor_speed_initial + position_x * max_speed_diff]
+    motor_speeds = [int(motor_speed_initial + position_x * max_speed_diff), int(motor_speed_initial - position_x * max_speed_diff), int(motor_speed_initial - position_x * max_speed_diff), int(motor_speed_initial + position_x * max_speed_diff)]
 
 def capture_image(server_url):
     response = requests.get(f"{server_url}/capture")
